GPS-21: Refactor DebugPanel and add NeedsNav component
This update modularizes the debug panel structure and introduces a fixed right-side needs navigation for improved user experience.

## Debug Panel Refactoring:
- **Component separation**: Extracted debug sections into individual components for better maintainability
  - `HungerSystemDebug.vue` - Hunger system controls and status
  - `PoopSystemDebug.vue` - Poop system timing configuration
  - `NeedsQueueDebug.vue` - Needs queue display using BoardList
  - `GameActionsDebug.vue` - Game reset and management actions
- **Cleaner architecture**: DebugPanel now orchestrates child components instead of managing all logic
- **Improved maintainability**: Each system's debug controls are self-contained with their own state management

## NeedsNav Component:
- **Fixed right-side navigation**: Always accessible toggle button at 50% screen height
- **BoardList integration**: Displays current needs with urgency indicators
- **Smooth interactions**: Simple fade-in/out animation without jarring slide effects
- **Responsive design**: Adapts panel width for different screen sizes
- **Non-intrusive overlay**: Doesn't interfere with main app layout

## UI/UX Improvements:
- **Decimal value cleanup**: Rounded hunger values and urgency percentages for cleaner display
- **BoardList consistency**: NeedsQueueDisplay now uses BoardList component for consistent styling
- **SubNav scroll fix**: Removed scroll prevention for better navigation experience
- **Mobile-first approach**: Statistics section always visible without Details wrapper

## Technical Enhancements:
- **Code organization**: Better separation of concerns across debug components
- **Reusable components**: BoardList used consistently across needs displays
- **Performance**: Reduced component complexity and improved maintainability
- **Accessibility**: Proper ARIA labels and semantic structure
